+++ include/device.h
@@ -27,6 +27,7 @@
  */
 
 #include <init.h>
+#include <pm/device.h>
 #include <sys/device_mmio.h>
 #include <sys/util.h>
 
@@ -332,39 +333,6 @@ typedef int16_t device_handle_t;
  */
 #define DEVICE_DECLARE(name) static const struct device DEVICE_NAME_GET(name)
 
-typedef void (*device_pm_cb)(const struct device *dev,
-			     int status, void *context, void *arg);
-
-/**
- * @brief Device PM info
- */
-struct device_pm {
-	/** Pointer to the device */
-	const struct device *dev;
-	/** Lock to synchronize the get/put operations */
-	struct k_sem lock;
-	/* Following are packed fields protected by #lock. */
-	/** Device pm enable flag */
-	bool enable : 1;
-	/* Following are packed fields accessed with atomic bit operations. */
-	atomic_t atomic_flags;
-	/** Device usage count */
-	atomic_t usage;
-	/** Device idle internal power state */
-	atomic_t fsm_state;
-	/** Work object for asynchronous calls */
-	struct k_work work;
-	/** Event object to listen to the sync request events */
-	struct k_poll_event event;
-	/** Signal to notify the Async API callers */
-	struct k_poll_signal signal;
-};
-
-/** Bit position in device_pm::atomic_flags that records whether the
- * device is busy.
- */
-#define DEVICE_PM_ATOMIC_FLAGS_BUSY_BIT 0
-
 /**
  * @brief Runtime device dynamic structure (in RAM) per driver instance
  *
@@ -650,77 +618,6 @@ static inline bool device_is_ready(const struct device *dev)
  */
 
 /**
- * @brief Device Power Management APIs
- * @defgroup device_power_management_api Device Power Management APIs
- * @ingroup power_management_api
- * @{
- */
-
-#ifdef CONFIG_PM_DEVICE
-
-/** @def DEVICE_PM_ACTIVE_STATE
- *
- * @brief device is in ACTIVE power state
- *
- * @details Normal operation of the device. All device context is retained.
- */
-#define DEVICE_PM_ACTIVE_STATE          1
-
-/** @def DEVICE_PM_LOW_POWER_STATE
- *
- * @brief device is in LOW power state
- *
- * @details Device context is preserved by the HW and need not be
- * restored by the driver.
- */
-#define DEVICE_PM_LOW_POWER_STATE       2
-
-/** @def DEVICE_PM_SUSPEND_STATE
- *
- * @brief device is in SUSPEND power state
- *
- * @details Most device context is lost by the hardware.
- * Device drivers must save and restore or reinitialize any context
- * lost by the hardware
- */
-#define DEVICE_PM_SUSPEND_STATE         3
-
-/** @def DEVICE_PM_FORCE_SUSPEND_STATE
- *
- * @brief device is in force SUSPEND power state
- *
- * @details Driver puts the device in suspended state after
- * completing the ongoing transactions and will not process any
- * queued work or will not take any new requests for processing.
- * Most device context is lost by the hardware. Device drivers must
- * save and restore or reinitialize any context lost by the hardware.
- */
-#define DEVICE_PM_FORCE_SUSPEND_STATE	4
-
-/** @def DEVICE_PM_OFF_STATE
- *
- * @brief device is in OFF power state
- *
- * @details - Power has been fully removed from the device.
- * The device context is lost when this state is entered, so the OS
- * software will reinitialize the device when powering it back on
- */
-#define DEVICE_PM_OFF_STATE             5
-
-/* Constants defining support device power commands */
-#define DEVICE_PM_SET_POWER_STATE       1
-#define DEVICE_PM_GET_POWER_STATE       2
-
-#endif /* CONFIG_PM_DEVICE */
-
-/**
- * @brief Get name of device PM state
- *
- * @param state State id which name should be returned
- */
-const char *device_pm_state_str(uint32_t state);
-
-/**
  * @brief Indicate that the device is in the middle of a transaction
  *
  * Called by a device driver to indicate that it is in the middle of a
@@ -818,112 +715,7 @@ int device_any_busy_check(void);
  */
 int device_busy_check(const struct device *chk_dev);
 
-#ifdef CONFIG_PM_DEVICE_IDLE
-
-/* Device PM states */
-enum device_pm_state {
-	DEVICE_PM_STATE_ACTIVE = 1,
-	DEVICE_PM_STATE_SUSPENDED,
-	DEVICE_PM_STATE_SUSPENDING,
-	DEVICE_PM_STATE_RESUMING,
-};
-
-/**
- * @brief Enable device idle PM
- *
- * Called by a device driver to enable device idle power management.
- * The device might be asynchronously suspended if Idle PM is enabled
- * when the device is not use.
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- */
-void device_pm_enable(const struct device *dev);
-
-/**
- * @brief Disable device idle PM
- *
- * Called by a device driver to disable device idle power management.
- * The device might be asynchronously resumed if Idle PM is disabled
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- */
-void device_pm_disable(const struct device *dev);
-
-/**
- * @brief Call device resume asynchronously based on usage count
- *
- * Called by a device driver to mark the device as being used.
- * This API will asynchronously bring the device to resume state
- * if it not already in active state.
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- * @retval 0 If successfully queued the Async request. If queued,
- * the caller need to wait on the poll event linked to device
- * pm signal mechanism to know the completion of resume operation.
- * @retval Errno Negative errno code if failure.
- */
-int device_pm_get(const struct device *dev);
-
-/**
- * @brief Call device resume synchronously based on usage count
- *
- * Called by a device driver to mark the device as being used. It
- * will bring up or resume the device if it is in suspended state
- * based on the device usage count. This call is blocked until the
- * device PM state is changed to resume.
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- * @retval 0 If successful.
- * @retval Errno Negative errno code if failure.
- */
-int device_pm_get_sync(const struct device *dev);
-
-/**
- * @brief Call device suspend asynchronously based on usage count
- *
- * Called by a device driver to mark the device as being released.
- * This API asynchronously put the device to suspend state if
- * it not already in suspended state.
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- * @retval 0 If successfully queued the Async request. If queued,
- * the caller need to wait on the poll event linked to device pm
- * signal mechanism to know the completion of suspend operation.
- * @retval Errno Negative errno code if failure.
- */
-int device_pm_put(const struct device *dev);
-
-/**
- * @brief Call device suspend synchronously based on usage count
- *
- * Called by a device driver to mark the device as being released. It
- * will put the device to suspended state if is is in active state
- * based on the device usage count. This call is blocked until the
- * device PM state is changed to resume.
- *
- * @param dev Pointer to device structure of the specific device driver
- * the caller is interested in.
- * @retval 0 If successful.
- * @retval Errno Negative errno code if failure.
- */
-int device_pm_put_sync(const struct device *dev);
-#else
-static inline void device_pm_enable(const struct device *dev) { }
-static inline void device_pm_disable(const struct device *dev) { }
-static inline int device_pm_get(const struct device *dev) { return -ENOTSUP; }
-static inline int device_pm_get_sync(const struct device *dev) { return -ENOTSUP; }
-static inline int device_pm_put(const struct device *dev) { return -ENOTSUP; }
-static inline int device_pm_put_sync(const struct device *dev) { return -ENOTSUP; }
 #endif
-#endif
-
-/** Alias for legacy use of device_pm_control_nop */
-#define device_pm_control_nop __DEPRECATED_MACRO NULL
 
 /**
  * @}
