+++ subsys/power/device_pm.c
@@ -8,6 +8,7 @@
 #include <kernel.h>
 #include <device.h>
 #include <sys/__assert.h>
+#include <pm/device_runtime.h>
 
 #define LOG_LEVEL CONFIG_PM_LOG_LEVEL /* From power module Kconfig */
 #include <logging/log.h>
