+++ subsys/power/power.c
@@ -9,8 +9,8 @@
 #include <timeout_q.h>
 #include <init.h>
 #include <string.h>
-#include <power/power.h>
-#include <power/power_state.h>
+#include <pm/pm.h>
+#include <pm/state.h>
 #include "policy/pm_policy.h"
 
 #define PM_STATES_LEN (1 + PM_STATE_SOFT_OFF - PM_STATE_ACTIVE)
